[2024-04-28] Implement live video streaming and object detection
This update includes the addition of live video streaming functionality to the application through the creation of a new `CameraStreaming` service. Furthermore, this service uses the YOLO model for object detection and draws bounding boxes on detected objects in each video frame. New endpoints were also added to support this feature.